move APInt utility functions
