+++ panda/plugins/taint2/taint_utils.h
@@ -1,26 +1,21 @@
 #ifndef __TAINT_UTILS_H_
 #define __TAINT_UTILS_H_
 
-#include "qemu/host-utils.h"
+#include <llvm/ADT/APInt.h>
 
-/**
- * Count leading zeroes for unsigned 128-bit integers.
- */
-static inline int clz128(unsigned __int128 val)
+static inline uint64_t apint_hi_bits(llvm::APInt value)
 {
-    uint64_t hi = static_cast<uint64_t>(val >> 64);
-    uint64_t lo = static_cast<uint64_t>(val);
-    return 0 != hi ? clz64(hi) : clz64(hi) + clz64(lo);
+    return value.lshr(64).trunc(64).getZExtValue();
 }
 
-/**
- * Count trailing zeroes for unsigned 128-bit integers.
- */
-static inline int ctz128(unsigned __int128 val)
+static inline uint64_t apint_lo_bits(llvm::APInt value)
 {
-    uint64_t hi = static_cast<uint64_t>(val >> 64);
-    uint64_t lo = static_cast<uint64_t>(val);
-    return 0 != lo ? ctz64(lo) : ctz64(hi) + ctz64(lo);
+    return value.trunc(64).getZExtValue();
+}
+
+static inline llvm::APInt make_128bit_apint(uint64_t hi, uint64_t lo)
+{
+    return (llvm::APInt(128, hi) << 64) | llvm::APInt(128, lo);
 }
 
 #endif
