+++ panda/plugins/taint2/tests/update_cb_switch/update_cb_switch.cpp
@@ -28,21 +28,6 @@
 
 const llvm::APInt NOT_LITERAL(128, ~0UL);
 
-static inline uint64_t apint_hi_bits(llvm::APInt value)
-{
-    return value.lshr(64).trunc(64).getZExtValue();
-}
-
-static inline uint64_t apint_lo_bits(llvm::APInt value)
-{
-    return value.trunc(64).getZExtValue();
-}
-
-static inline llvm::APInt make_128bit_apint(uint64_t hi, uint64_t lo)
-{
-    return (llvm::APInt(128, hi) << 64) | llvm::APInt(128, lo);
-}
-
 /*
  * Run a test, and print out the results.  Note that not all arguments are used
  * by all tests, and may be dummied up for those tests.
